--- src/components/signCard/index.jsx
@@ -1,12 +1,0 @@
-import { Button, Text, View } from "react-native";
-
-const signCard = () => {
-	return (
-		<View>
-			<Text>HomePage</Text>
-      <Button onPress={()=> }></Button>
-		</View>
-	);
-};
-
-export default Home;
