+++ src/pages/Login/index.jsx
@@ -1,8 +1,23 @@
-export default function Login() {
+import { View } from "react-native";
+import BackgroundImage from "../../components/BackgroundImage";
+
+ const Login = () => {
+
 	return (
-		<View>
-			<Text>LOGIN PAGE</Text>
-			<Button></Button>
-		</View>
+    <BackgroundImage>
+        <View style={styles.container}>
+            <Slogan text="A melhor plataforma para prestação e contratação de serviços" />
+            <View style={styles.SignContainer}>
+
+            </View>
+            <View style={styles.buttonAlign}>
+              <BirdsButton title="Trabalhar" onPress={} />
+              <BirdsButton title="Contratar" contractButton={true} />
+          </View>
+        </View>
+
+    </BackgroundImage>
 	);
 }
+
+export default Login;
