+++ src/routes/app.routes.jsx
@@ -1,13 +1,14 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import Home from "../components/signCard";
 import Login from "../pages/Login";
+import PerfilChoose from "../pages/PerfilChoose";
+
 const { Navigator, Screen } = createNativeStackNavigator();
 
 export function AppRoutes() {
 	return (
 		<Navigator>
-			<Screen name="home" component={Home} />
-			<Screen name="Login" component={Login} />
+			<Screen name="perfilChoose" component={PerfilChoose} />
+			<Screen name="login" component={Login} />
 		</Navigator>
 	);
 }
