+++ src/routes/index.jsx
@@ -1,4 +1,4 @@
-import { NavigationContainer } from "react-navigation/native";
+import { NavigationContainer } from "@react-navigation/native";
 import { AppRoutes } from "./app.routes";
 
 export function Routes() {
